chore: add wakeup loading page, move text
