+++ client/src/App.tsx
@@ -5,12 +5,14 @@ import { Createpage } from './Pages/Createpage'
 import { Titlepage } from './Pages/Titlepage'
 import { RegisterForm } from './Pages/Register'
 import { LoginForm } from './Pages/Login'
+import { WakeUp } from './Pages/WakeUpPage'
 
 function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Titlepage />} />
+        <Route path="/" element={<WakeUp />} />
+        <Route path="/home" element={<Titlepage />} />
         <Route path="/recipes" element={<Homepage />} />
         <Route path="/recipes/:id" element={<Recipepage />} />
         <Route path="/add-recipe" element={<Createpage />} />
