+++ client/src/Pages/Titlepage.tsx
@@ -1,86 +1,12 @@
-import {
-  Box,
-  ButtonGroup,
-  Center,
-  Flex,
-  Heading,
-  Image,
-  Spinner,
-  Text,
-  useToast,
-} from '@chakra-ui/react'
+import { Box, ButtonGroup, Flex, Heading, Image, Text } from '@chakra-ui/react'
 import { Header } from '../Components/Shared/Header/Header'
 import { Footer } from '../Components/Shared/Footer/Footer'
 import potImage from '../assets/onder-ortel-rpMvlIvT9hM-unsplash.jpg'
 import { CornerRightDown } from 'lucide-react'
 import { LoginRedirect, RegisterRedirect } from '../Components/Shared/Buttons/Button'
 import { Link } from 'react-router-dom'
-import axios from 'axios'
-import { useEffect, useState } from 'react'
-
-const THREE_THOUSAND = 3000
-const MAX_LOADING_DURATION = 10000
-
-const loadingMessages = [
-  'Server se spouští... asi. Možná. 😬',
-  'Server si dal šlofíčka. Vydržte chvilku. 😪',
-  'Server má právě kreativní pauzu, vydržte chvilku. ✨',
-  'Server si dopřává trochu kávy. ☕',
-]
 
 export function Titlepage() {
-  const toast = useToast()
-  const [loading, setLoading] = useState(true)
-  const [loadingMessage, setLoadingMessage] = useState('')
-
-  useEffect(() => {
-    setLoadingMessage(loadingMessages[Math.floor(Math.random() * loadingMessages.length)])
-
-    const fetchData = async () => {
-      const timeoutId = setTimeout(() => {
-        setLoading(false)
-        toast({
-          title: 'Časový limit serveru',
-          description: 'Server reaguje příliš dlouho. Zkuste to znovu později.',
-          status: 'error',
-          duration: THREE_THOUSAND,
-          isClosable: true,
-        })
-      }, MAX_LOADING_DURATION)
-
-      try {
-        const response = await axios.get('/wakeUp')
-        if (response.status === 200) {
-          setLoading(false)
-          clearTimeout(timeoutId)
-        }
-      } catch (error) {
-        setLoading(false)
-        clearTimeout(timeoutId)
-        toast({
-          title: 'Chyba',
-          description: 'Nepodařilo se probudit server. Zkuste to znovu později.',
-          status: 'error',
-          duration: THREE_THOUSAND,
-          isClosable: true,
-        })
-      }
-    }
-
-    fetchData()
-  }, [toast])
-
-  if (loading) {
-    return (
-      <Center h="full" flexDirection="column" height="100vh">
-        <Spinner color="teal.500" size="lg" borderWidth="4px" />
-        <Text mt="20px" fontSize="25px">
-          {loadingMessage}
-        </Text>
-      </Center>
-    )
-  }
-
   return (
     <Box bg="#d0ffd5">
       <Header />
@@ -113,7 +39,7 @@ export function Titlepage() {
                 },
                 '@media screen and (max-width: 509px)': {
                   fontSize: '35px',
-                  ml: '25px',
+                  ml: '8px',
                   mb: '50px',
                 },
               }}
@@ -132,7 +58,7 @@ export function Titlepage() {
                 },
                 '@media screen and (max-width: 509px)': {
                   fontSize: '15px',
-                  ml: '25px',
+                  ml: '10px',
                   mb: '45px',
                 },
               }}
@@ -151,7 +77,7 @@ export function Titlepage() {
                 },
                 '@media screen and (max-width: 509px)': {
                   fontSize: '15px',
-                  ml: '25px',
+                  ml: '10px',
                   mb: '45px',
                 },
               }}
@@ -162,7 +88,7 @@ export function Titlepage() {
             <Flex
               direction="row"
               my="30px"
-              ml="35px"
+              ml="42px"
               fontSize="19px"
               sx={{
                 '@media screen and (max-width: 699px)': {
@@ -172,14 +98,14 @@ export function Titlepage() {
                 },
                 '@media screen and (max-width: 509px)': {
                   fontSize: '15px',
-                  ml: '25px',
+                  ml: '10px',
                   mb: '45px',
                 },
               }}
             >
               <Text>Nemáte účet? Zaregistrujte se zde</Text>
-              <Box>
-                <CornerRightDown />
+              <Box ml="5px" mt="10px">
+                <CornerRightDown size="20px" />
               </Box>
             </Flex>
             <ButtonGroup
@@ -192,8 +118,9 @@ export function Titlepage() {
                 },
                 '@media screen and (max-width: 509px)': {
                   fontSize: '15px',
-                  ml: '90px',
+                  ml: '10px',
                   width: '250px',
+                  flexDirection: 'column',
                 },
               }}
             >
